Get discrete testing passing and begin work on repairing continuous
